% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{checkpointer}
\alias{checkpointer}
\title{Create a workflow checkpointer}
\usage{
checkpointer(backend = c("memory", "file"), path = NULL)
}
\arguments{
\item{backend}{Either \code{"memory"} (in-process) or \code{"file"} (directory of JSON
files).}

\item{path}{Directory path for the file backend.}
}
\value{
A \code{Checkpointer} R6 object.
}
\description{
Persists graph execution state so workflows can be resumed.
}
\examples{
cp <- checkpointer()
cp$save("thread-1", "node_a", list(x = 1))
cp$load("thread-1")
}
