% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestr-package.R
\docType{data}
\name{END}
\alias{END}
\title{End sentinel for graph execution}
\format{
An object of class \code{character} of length 1.
}
\usage{
END
}
\description{
Use \code{END} as a target node name to indicate that graph execution should
stop after the current node.
}
\examples{
# Use END as the target of a graph edge to stop execution
END
g <- graph_builder()
g$add_node("a", function(state, config) list(x = 1))
g$add_edge("a", END)
}
\seealso{
Other graph-building: 
\code{\link{graph_builder}()},
\code{\link{state_schema}()}
}
\concept{graph-building}
\keyword{datasets}
