% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-helpers.R
\name{route_to}
\alias{route_to}
\title{Create a constant router}
\usage{
route_to(node_name)
}
\arguments{
\item{node_name}{Character node name to route to.}
}
\value{
A function that always returns the given node name.
}
\description{
Returns a condition function that always routes to the given node name.
}
