% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestr-package.R
\docType{data}
\name{END}
\alias{END}
\title{End sentinel for graph execution}
\format{
An object of class \code{character} of length 1.
}
\usage{
END
}
\description{
Use \code{END} as a target node name to indicate that graph execution should
stop after the current node.
}
\keyword{datasets}
